Default core-lambda-stubs to build from source
Given that core-lambda-stubs compile filegroups that are already in the
stubs format and metalava is not in the critical path of
core-lambda-stubs, there is hardly any benefit in building
core-lambda-stubs from API signature files. Therefore, remove the
from-source vs from-text toggle for core-lambda-stubs and always build
the module from source (stubs).

Test: m nothing --build-from-text-stub
Bug: 287522618
Change-Id: Iea3f6c7e2f71914c97b209ae548f06a502bee60e