Work around a possible optimization.
clang for riscv64 currently does the loads for the comparisons on the
_right_ hand side of an `&&` before doing the comparsion on the left
hand side. This is a problem for this particular code, because we
don't know whether the structure is large enough until after we've done
the check on the left hand side.

This doesn't matter on Android, because bionic will always calloc() a
full sockaddr_storage for anything it gets back from netd (so the worst
that can happen is we read a few zeros).

This is the minimal workaround.

Bug: https://github.com/google/android-riscv64/issues/124
Test: treehugger
Change-Id: I52eebef07fa67c001f10843001a1c57570bda92a