Pass sockaddr rather than sockaddr_storage to sockaddrToInetAddress.
Although on Android, bionic will always calloc() a sockaddr_storage for
each result from netd, on the host (including with host bionic) you
might actually only have a sockaddr_in. Normally this isn't problematic,
because the `&&` means we don't look at anything beyond the first
"family" element of the struct without checking the family first, which
tells us how large the object really is.

For historical microcontroller reasons, when targeting riscv64 clang is
so desperate to avoid branches that it takes advantage of the fact that
we've claimed to have a sockaddr_storage (which is "at least as large"
as any of the actual sockaddr structs) to remove the `&&`. Which works
fine for addresses from netd, but not necessarily on the host.

So... stop lying to the compiler. Until we look at the family, we don't
actually know (except when we're on a device) how big the struct is, so
let's not claim anything more than "it's a sockaddr".

Idependently, we should tell clang not to do this, because it's almost
certainly a bad choice for non-microcontrollers!

Idependently, we should probably work out how to delete this function
and use the near-identical (but already safe against this transformation)
NET_SockaddrToInetAddress.

Bug: https://github.com/google/android-riscv64/issues/124
Test: treehugger
Change-Id: Ibd4ee9c1f19c40fdac43142d244d59ab89bbd169