Introduce @FlaggedApi to libcore/
Until go/aconfig-unbundled is implemented, @FlaggedApi APIs should not
be used by Android platform code, except used by ART and libcore itself,
because the -next build target is still using ART 14 module.

The annotation is only used to generate current.txt and stub files.

java_aconfig_library will be added for the flag in the short term:
The java_aconfig_library isn't part of core-all, core-art or core-oj
and not in the bootclasspath right now, until reading the flag value
from the module is actually supported.

The java_aconfig_library shouldn't be visible to the frameworks either.
Until go/aconfig-unbundled is supported, frameworks should not use the
APIs. When Android Linter forces checking flag (See b 303434307 for
details), it should make it harder for frameworks to use the APIs until
we make the aconfig library visible.

Bug: 313852508
Test: m droid
Change-Id: Ide26c2c75828ac2665521b1491b9cdfc735aa274