Use the generated aconfig constant to flag APIs in luni/
Ideally, we can annotate luni/ APIs with @FlaggedApi. However,
The libcore_aconfig_flags_lib produces a compiled jar, not .java file
sources. It's not a genrule or gensrcs module, and thus it depends
core-all and causes a circular dependency.

To avoid this circular build dependency, we only annotate the api stubs,
not the java source files directly.

@FlaggedApi retents in RetentionPolicy.SOURCE only, and it shouldn't
affect the runtime.

Bug: 316554963
Test: m droid
Change-Id: I6632a7aa230f603ed5612c15fcbb3463523f5d3d