Offer `core-libart-for-host` under Ravenwood.
Ravenwood is a deviceless testing environment that runs in a host
JVM, similar to Robolectric.

As we're working on onboard existing SystemUI tests that expect
parity with Robolectric, we're needing to offer the same baseline
Robolectric functionality, so we expand the visibility of
`core-libart-for-host` to only the top-level of "//frameworks/base"
which is where Ravenwood.bp is defined.  (We're not exposing the
library to any subpackages.)

Bug: 319647875
Test: atest SystemUiRavenTests
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:ec696e648834239199fb0f26500f5b95d8fa0a6b)
Merged-In: Ibeb488ef1d4e3cb59104c37e269b7212b7ca7f20
Change-Id: Ibeb488ef1d4e3cb59104c37e269b7212b7ca7f20