Move LinkedHashMapTest.test_entryCompatibility_* tests
We have imported LinkedHashMap.Entry class from the upstream
to make sure LinkedHashMap.Entry refers to the same class in
build-time in libcore/, i.e. LinkedHashMap.Entry.
It's more consistent with the upstream code.

For apps, LinkedHashMap.Entry still refers to Map.Entry as
the nested LinkedHashMap.Entry class is not in the Android SDK.

Moving the test to build against module_current SDK.

Bug: 320460092
Test: atest libcore-sdk-current-test
Change-Id: Ic1b67a47ce7f1a9f07c9d87f87cf9bb50c2086cf