Regenerate expired test certificates.
Minor test updates to handle SHA1 deprecation so
signatures are all SHA256 based now.

The regenerated test cert for getSigAlgParams() causes
failures due to extra information in the new cert
not matching the hand coded ASN.1 in the test, but
it passes with the expired so will redo later.

This will need backporting to a bunch of branches....

Bug: 322550827
Test: atest CtsLibcoreTestCases:libcore.java.security.cert
Change-Id: I78b737bb05678b155e17aeeca19f5319437f9353