Ignore AKISerialNumberTest.
This is a test for a corner case that doesn't often happen in
practice and for which the test certificates have expired.
It will take a bit of time to generate new ones so I am
disabling the test for now, will re-enable it on main etc
early next week.

Needs cherry-picking to many branches to fix CI.

Bug: 323360381
Test: atest CtsLibcoreOjTestCases:test.java.security.cert.AKISerialNumberTest
Change-Id: Icf51069c27d79557ec2dfa6462da1a557a717e48